+++ nossaBiblioteca/src/main/java/br/com/fafic/ppi/nossaBiblioteca/service/BibliotecaService.java
@@ -3,6 +3,7 @@ package br.com.fafic.ppi.nossaBiblioteca.service;
 import br.com.fafic.ppi.nossaBiblioteca.domain.Biblioteca;
 import br.com.fafic.ppi.nossaBiblioteca.domain.Bibliotecario;
 import br.com.fafic.ppi.nossaBiblioteca.domain.exception.ObjectNotFoundException;
+import br.com.fafic.ppi.nossaBiblioteca.dto.BibliotecaDTO;
 import br.com.fafic.ppi.nossaBiblioteca.repositories.BibliotecaRepository;
 import br.com.fafic.ppi.nossaBiblioteca.repositories.BibliotecarioRepository;
 import lombok.RequiredArgsConstructor;
@@ -16,7 +17,12 @@ public class BibliotecaService {
 
     private final BibliotecaRepository bibliotecaRepository;
 
-    public Biblioteca save(Biblioteca biblioteca){
+    public Biblioteca save(BibliotecaDTO bibliotecaDTO){
+        var biblioteca = new Biblioteca(bibliotecaDTO.getNome(),
+                bibliotecaDTO.getBibliotecario(),
+                bibliotecaDTO.getLivros(),
+                bibliotecaDTO.getAlunos(),
+                bibliotecaDTO.getProfessores());
         return bibliotecaRepository.save(biblioteca);
     }
 
@@ -29,6 +35,12 @@ public class BibliotecaService {
                 .orElseThrow(()-> new ObjectNotFoundException("Não existe na base de dados o Id = " + id));
     }
     public void deleteBiblioteca(Long id) {
-        bibliotecaRepository.deleteById(id);
+        var biblioteca = bibliotecaRepository.findById(id);
+        if(biblioteca.isPresent()){
+            bibliotecaRepository.deleteById(id);
+        }
+        else {
+            biblioteca.orElseThrow(()-> new ObjectNotFoundException("Não existe na base de dados o Id = "+ id));
+        }
     }
 }
