+++ nossaBiblioteca/src/main/java/br/com/fafic/ppi/nossaBiblioteca/service/DevolucaoService.java
@@ -3,6 +3,7 @@ package br.com.fafic.ppi.nossaBiblioteca.service;
 import br.com.fafic.ppi.nossaBiblioteca.domain.Devolucao;
 import br.com.fafic.ppi.nossaBiblioteca.domain.Emprestimo;
 import br.com.fafic.ppi.nossaBiblioteca.domain.exception.ObjectNotFoundException;
+import br.com.fafic.ppi.nossaBiblioteca.dto.DevolucaoDTO;
 import br.com.fafic.ppi.nossaBiblioteca.repositories.DevolucaoRepository;
 import br.com.fafic.ppi.nossaBiblioteca.repositories.EmprestimoRepository;
 import lombok.RequiredArgsConstructor;
@@ -16,7 +17,9 @@ public class DevolucaoService {
 
     private final DevolucaoRepository devolucaoRepository;
 
-    public Devolucao save(Devolucao devolucao){
+    public Devolucao save(DevolucaoDTO devolucaoDTO){
+        var devolucao = new Devolucao(devolucaoDTO.getEmprestimo(),
+                devolucaoDTO.getDataDeDevolucao());
         return devolucaoRepository.save(devolucao);
     }
 
@@ -29,6 +32,12 @@ public class DevolucaoService {
                 .orElseThrow(()-> new ObjectNotFoundException("Não existe na base de dados o Id = " + id));
     }
     public void deleteDevolucao(Long id) {
-        devolucaoRepository.deleteById(id);
+        var devolucao = devolucaoRepository.findById(id);
+        if(devolucao.isPresent()){
+            devolucaoRepository.deleteById(id);
+        }
+        else {
+            devolucao.orElseThrow(()-> new ObjectNotFoundException("Não existe na base de dados o Id = "+ id));
+        }
     }
 }
