+++ nossaBiblioteca/src/main/java/br/com/fafic/ppi/nossaBiblioteca/service/EmprestimoService.java
@@ -1,8 +1,10 @@
 package br.com.fafic.ppi.nossaBiblioteca.service;
 
 import br.com.fafic.ppi.nossaBiblioteca.domain.Emprestimo;
+import br.com.fafic.ppi.nossaBiblioteca.domain.Livro;
 import br.com.fafic.ppi.nossaBiblioteca.domain.Login;
 import br.com.fafic.ppi.nossaBiblioteca.domain.exception.ObjectNotFoundException;
+import br.com.fafic.ppi.nossaBiblioteca.dto.EmprestimoDTO;
 import br.com.fafic.ppi.nossaBiblioteca.repositories.EmprestimoRepository;
 import br.com.fafic.ppi.nossaBiblioteca.repositories.LoginRepository;
 import lombok.RequiredArgsConstructor;
@@ -15,8 +17,13 @@ import java.util.List;
 public class EmprestimoService {
 
     private final EmprestimoRepository emprestimoRepository;
+    private Livro livro;
 
-    public Emprestimo save(Emprestimo emprestimo){
+    public Emprestimo save(EmprestimoDTO emprestimoDTO){
+        var emprestimo = new Emprestimo(emprestimoDTO.getAluno(),
+                emprestimoDTO.getProfessor(),
+                emprestimoDTO.getDataDoEmprestimo(),
+                emprestimoDTO.getLivro());
         return emprestimoRepository.save(emprestimo);
     }
 
@@ -29,6 +36,12 @@ public class EmprestimoService {
                 .orElseThrow(()-> new ObjectNotFoundException("Não existe na base de dados o Id = " + id));
     }
     public void deleteEmprestimo(Long id) {
-        emprestimoRepository.deleteById(id);
+        var empretimo = emprestimoRepository.findById(id);
+        if(empretimo.isPresent()){
+            emprestimoRepository.deleteById(id);
+        }
+        else {
+            empretimo.orElseThrow(()-> new ObjectNotFoundException("Não existe na base de dados o Id = "+ id));
+        }
     }
 }
