+++ nossaBiblioteca/src/main/java/br/com/fafic/ppi/nossaBiblioteca/service/LivroService.java
@@ -2,6 +2,7 @@ package br.com.fafic.ppi.nossaBiblioteca.service;
 
 import br.com.fafic.ppi.nossaBiblioteca.domain.Livro;
 import br.com.fafic.ppi.nossaBiblioteca.domain.exception.ObjectNotFoundException;
+import br.com.fafic.ppi.nossaBiblioteca.dto.LivroDTO;
 import br.com.fafic.ppi.nossaBiblioteca.repositories.LivroRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,7 +16,11 @@ public class LivroService {
 
     private final LivroRepository livroRepository;
 
-    public Livro save(Livro livro){
+    public Livro save(LivroDTO livroDTO){
+        var livro = new Livro(livroDTO.getIsbn(),
+                livroDTO.getNome(),
+                livroDTO.getArea(),
+                livroDTO.getEmprestimo());
         return livroRepository.save(livro);
     }
 
@@ -28,7 +33,13 @@ public class LivroService {
                 .orElseThrow(()-> new ObjectNotFoundException("Não existe na base de dados o Id = " + id));
     }
     public void deleteLivro(Long id) {
-        livroRepository.deleteById(id);
+        var livro = livroRepository.findById(id);
+        if(livro.isPresent()){
+            livroRepository.deleteById(id);
+        }
+        else {
+            livro.orElseThrow(()-> new ObjectNotFoundException("Não existe na base de dados o Id = "+ id));
+        }
     }
